requirements.txt
requests==2.28.2
psycopg2-binary
